Introduce a shortcode to manage donor honor rolls
We'll provide a large text input area at first to allow the copy
paste of a large list of names. We then convert this to a display
on the front end.

This is a shortcode that can be embedded in any page, but is more
easily managed via Shortcode UI